Move trait resolution to its own module
